Final Commit

+++ Front-End/src/js/main.js
@@ -1,5 +1,12 @@
-        let ul=document.querySelector('ul');
-        
+        let thead=document.querySelector('#thead');
+        let tbody=document.querySelector('#tbody');
+        // div Contents
+        let divlogin=document.getElementById('login');
+        let tbGame=document.getElementById('tabgame');
+        let divTabCriarGame=document.getElementById('divTabCriarGame');
+        let divGameNovo=document.querySelector('.cadastroNovoGame');
+        let VerList=document.getElementById('vertabela');
+        let editarGame=document.querySelector('.editarGame');
         let form1=document.getElementById('form1');
 
         let form2=document.getElementById('form2');
@@ -13,6 +20,16 @@
             }
         }
 
+        VerList.addEventListener('click',()=>{
+            ListGame();
+        });
+        
+        divTabCriarGame.addEventListener('click',()=>{
+            tbGame.classList.remove('classShow');
+            tbGame.classList.add('classHiden');
+            divGameNovo.classList.remove('classHiden');
+            divGameNovo.classList.add('classShow');
+        });
 
 function Logar() {
             
@@ -30,45 +47,80 @@ function Logar() {
                 let token=res.data.token;
                 localStorage.setItem("token",token);
                 axiosConfig.headers.Authorization="Bearer "+localStorage.getItem("token");
-
+                divlogin.classList.add('classHiden');
+                divGameNovo.classList.remove('classHiden');
+                divGameNovo.classList.add('classShow');
             }).catch(error=>{
                 alert(error);
             })
         }
 
+   function ListGame(){
+    tbody.innerHTML="";
         axios.get('http://localhost:8080/games',axiosConfig).then(response=>{
-            games=response.data.games
+            games=response.data.games;
+           
             games.forEach(element => {
-                let item=document.createElement('li')
-                item.innerHTML=`<strong>ID</strong>:${element.id} - <strong>Title</strong>:${element.title} - <strong>Ano</strong>:${element.year} - <strong>Preço</strong>:${element.price} </strong> <br>`
-                item.setAttribute("data-id",element.id);
-                item.setAttribute("data-title",element.title);
-                item.setAttribute("data-year",element.year);
-                item.setAttribute("data-price",element.price);
+                let itembody=document.createElement('tr');
+                // tbody
+
+                itembody.innerHTML=`<td>${element.id}</td>`
+                itembody.innerHTML+=`<td>${element.title}</td>`
+                itembody.innerHTML+=`<td>${element.year}</td>`
+                itembody.innerHTML+=`<td>${element.price} - kz</td>`
+
+                itembody.setAttribute("data-id",element.id);
+                itembody.setAttribute("data-title",element.title);
+                itembody.setAttribute("data-year",element.year);
+                itembody.setAttribute("data-price",element.price);
 
                 let btn=document.createElement('button')
-            btn.innerHTML="Deletar"
+                btn.classList.add('bg-red-500');
+                btn.classList.add('px-2');
+                btn.classList.add('py-1');
+                btn.classList.add('text-white');
+                btn.classList.add('rounded-md');
+                btn.classList.add('mr-4');
+                btn.classList.add('mt-4');
+                btn.innerHTML="Deletar"
 
             let btnEdit=document.createElement('button')
             btnEdit.innerHTML="Editar"
+            btnEdit.classList.add('bg-blue-500');
+            btnEdit.classList.add('px-2');
+            btnEdit.classList.add('py-1');
+            btnEdit.classList.add('text-white');
+            btnEdit.classList.add('rounded-md');
+            btnEdit.classList.add('mr-4');
+            btnEdit.classList.add('mt-4');
+          
+            itembody.appendChild(btnEdit);
+            itembody.appendChild(btn);
+            
+            tbody.appendChild(itembody);
 
-            item.appendChild(btn);
-            item.appendChild(btnEdit);
-            ul.appendChild(item)
-
-            btn.addEventListener("click",()=>{
-                let id=item.getAttribute("data-id");
-                DeleteGame(id);
+            btn.addEventListener("click",(e)=>{
+                let id=itembody.getAttribute("data-id");
+                DeleteGame(id,e.target);
             })
             
 
-            btnEdit.addEventListener("click",()=>{
-                LoadListItem(item);
+            btnEdit.addEventListener("click",(e)=>{
+                LoadListItem(itembody);
             })
+
             });
+          
+            divGameNovo.classList.remove('classShow');
+            divGameNovo.classList.add('classHiden');
+            tbGame.classList.remove('classHiden');
+            tbGame.classList.add('classShow');
         }).catch(error=>{
             console.log(error)
         });
+    }
+    
+       
 
         form1.addEventListener('submit',(e)=>{
             e.preventDefault();
@@ -78,7 +130,7 @@ function Logar() {
 
         form2.addEventListener('submit',(e)=>{
             e.preventDefault();
-            UpdateGame();
+            UpdateGame(e.target);
             
         })
 
@@ -95,19 +147,24 @@ function Logar() {
             }
 
         axios.post('http://localhost:8080/game',game,axiosConfig).then(result=>{
-                alert(result.statusText);
+            title.value=null;
+            year.value=null;
+            price.value=null;
+            ListGame();
         }).catch(error=>{
-                console.log(error)
+        alert(error);
         })
             
         }
 
-        function DeleteGame(id){
-        axios.delete('http://localhost:8080/game/'+id,axiosConfig).then(result=>{
-            console.log(result)
-        }).catch(error=>{
-            console.log(error);
-        })
+        function DeleteGame(id,el){
+            if(confirm('Tem certeza que deseja eliminar este game? ')){
+                axios.delete('http://localhost:8080/game/'+id,axiosConfig).then(result=>{ 
+                    tbody.removeChild(el.parentNode);
+                }).catch(error=>{
+                    alert(error);
+                })
+            }
 
         }
 
@@ -121,7 +178,10 @@ function Logar() {
             title.value=item.getAttribute('data-title');
             year.value=item.getAttribute('data-year');
             price.value=item.getAttribute('data-price');
-
+            tbGame.classList.remove('classShow');
+            tbGame.classList.add('classHiden');
+            editarGame.classList.remove('classHiden');
+            editarGame.classList.add('classShow');
         }
 
         function UpdateGame(){
@@ -136,10 +196,18 @@ function Logar() {
                 year:year.value,
                 price:price.value
             }
-
+           
+           
             axios.put('http://localhost:8080/game/'+id.value,game,axiosConfig).then(result=>{
-                alert(result.statusText);
+                //alert(result.statusText);
+                ListGame();
+                //tbody.removeChild
         }).catch(error=>{
                 console.log(error)
         })
+        tbGame.classList.add('classShow');
+        tbGame.classList.remove('classHiden');
+        editarGame.classList.add('classHiden');
+        editarGame.classList.remove('classShow');
+
         }
